requirements.txt
black
pylint
pytest
flask
matplotlib
numpy
scipy
scikit-learn
pandas
pytest--cov
gunicorn
uvloop
httptools 
google-cloud-aiplatform